% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_objects.R
\docType{data}
\name{dir_data_prep}
\alias{dir_data_prep}
\title{Directory where data is processed}
\format{
An object of class \code{character} of length 1.
}
\usage{
dir_data_prep
}
\description{
Directory where data is processed
}
\keyword{datasets}
