% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-utils.R
\name{fortify_vars}
\alias{fortify_vars}
\title{Put variables into correct format}
\usage{
fortify_vars(iter, regex = "var_*", cn = NULL, default_ds = "data_raw")
}
\description{
Put variables into correct format
}
