% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_objects.R
\name{get_dir_data_prep}
\alias{get_dir_data_prep}
\title{Get directory where raw data are kept}
\usage{
get_dir_data_prep(dataset)
}
\arguments{
\item{dataset}{"cytof" or "antibody".}
}
\value{
Character.
}
\description{
Gives directory where raw data, such as plots, are kept.
Assumes that the source folder is kept in same directory as acsanalysis package,
and that the final folder in the directory is paste0(dataset, "acs").
}
