% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-utils.R
\name{get_var_clin}
\alias{get_var_clin}
\title{Get clinical variables to add to a dataset}
\usage{
get_var_clin(var, var_clin_possible)
}
\arguments{
\item{var}{character vector.
Variables to add not specificied by iter.}

\item{iter}{one-row dataframe.}

\item{iter_cn}{character vector.
Names of columns in
\code{iter} to get clinical variable names from.}
}
\description{
Get clinical variables to add to a dataset
}
