% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pkg_private.r
\name{install_pkg_private}
\alias{install_pkg_private}
\title{Install private GitHub package dependencies}
\usage{
install_pkg_private(
  project,
  update = FALSE,
  gitcreds_url = "https://github.com"
)
}
\arguments{
\item{project}{"acs_cytof_tcells", "acs_cytof_nkbcells", "antibody".
Determines which packages are required for installation.}

\item{update}{logical.
If \code{FALSE}, then a package would
not be installed if it is already installed,
even if an update were available.
Default is \code{FALSE}.}

\item{gitcreds_url}{URL connected with GitHub PAT.
Passed to url argument of
\code{gitcreds::gitcreds_get}.
Default is \code{"https://github.com"}}
}
\description{
Install private GitHub package dependencies
}
