% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-plot_disp_int_prog_num.R
\name{plot_disp_int_cat_num}
\alias{plot_disp_int_cat_num}
\title{Plot model est and ci for int between cat and num exp var}
\usage{
plot_disp_int_cat_num(
  mod,
  .data,
  data_nm,
  var_cat,
  var_num,
  var_offset,
  var_dep,
  max_sd = NULL,
  axis_x_reverse = FALSE,
  cat_to_col = NULL,
  axis_lab = NULL,
  add_test = "lr",
  trans = scales::identity_trans(),
  dir_test,
  table_coord = c(0.1, 1),
  table_size_skip = 0.05,
  table_size_text = 10,
  range_extend = 0,
  point_alpha = 1,
  limits_include = NULL,
  gg_theme = cowplot::theme_cowplot(),
  grid = "xy",
  point_size = NULL
)
}
\description{
Plot model est and ci for int between cat and num exp var
}
