% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-plot_display_auto.R
\name{plot_display_auto}
\alias{plot_display_auto}
\title{Make automatic plot using effects package}
\usage{
plot_display_auto(
  mod,
  data_mod,
  var,
  fn,
  dir_save,
  height = NULL,
  width = NULL
)
}
\arguments{
\item{mod}{model object.}

\item{data_mod}{dataframe.}

\item{var}{character vector. Names of variables to be saved. Only required
if height or width is NULL, as its length is used to calculate
them automatically.}

\item{fn}{character. File name. Default is \code{"p_fit_auto.png"}.}

\item{dir_save}{character. Folder to save to.}

\item{height, width}{numeric. If not \code{NULL}, then
the plot height and width are given by these.
Default is \code{NULL}.}
}
\value{
Returns path to save to invisibly.
}
\description{
Make automatic plot using effects package
}
\details{
Note that only PNG output is supported at present.
The plot function for `effects::predictorEffects` assumes
that the data used to fit the model
are in the global environment. This function
automatically assigns the data to the global environment
to an object with the correct name, and restores
any object "overwritten" after the function is completed running.
}
