% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-utils.R
\name{prep_bs_freq}
\alias{prep_bs_freq}
\title{Calculate bs-freq for CyTOF phenotypes}
\usage{
prep_bs_freq(.data, cyt_response_type)
}
\arguments{
\item{.data}{dataframe.}

\item{cyt_response_type}{"summed" or "combn". Specifies
type of response to create.
Should be able to specify "single" as well.}
}
\description{
Calculate bs-freq for CyTOF phenotypes
}
