% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iter.R
\name{print_iter}
\alias{print_iter}
\title{Print iterator}
\usage{
print_iter(iter, ind)
}
\arguments{
\item{ind}{integer. Row index of the iterator.}
}
\description{
Prints all factors
in current iterator that are different to the last.
Assumes that \code{iter} is the name of the iterator
}
