% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_effects_Effect.R
\name{run_effects_Effect}
\alias{run_effects_Effect}
\title{Run effects::effect when in an execution environment}
\usage{
run_effects_Effect(data, nm, var, mod)
}
\value{
output from \code{effects::Effect(focal.predictors = var, mod = mod, xlevels = 50)}.
}
\description{
Run effects::effect when in an execution environment
}
