% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-plot_disp_int_prog_num.R
\name{plot_disp_int_cat_num}
\alias{plot_disp_int_cat_num}
\title{Plot model est and ci for int between cat and num exp var}
\usage{
plot_disp_int_cat_num(
  mod,
  .data,
  data_nm,
  var_cat,
  var_num,
  var_offset,
  var_dep,
  axis_x_reverse = FALSE,
  cat_to_col = NULL,
  axis_lab = NULL,
  add_test = "lr",
  dir_test,
  gg_theme = cowplot::theme_cowplot(),
  grid = "xy",
  point_size = NULL
)
}
\description{
Plot model est and ci for int between cat and num exp var
}
